+++ packages/widget-dashboard/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.6.0 (2026-08-26)
+
 ### New Features
 
 -   Widget chrome footer: `relevance: 'high'` actions mount as leading text